take the losses out
